1. fix time.Time to msgpack.spec    a. reade_bytes.go  -> ReadTimeBytes    b. write_bytes.go  -> AppendTime    c. simple_marshal    d. simple_unmarshal 1. change TimeExtension from 5 -> 255(-1) 2.    var extensionReg = make(map[int8]func() Extension)    -> var extensionReg = make(map[uint8]func() Extension) 3. type ExtensionTypeError struct { 	Got  int8  // -> uint8 	Want int8  // -> uint8     } 4. type Extension interface { 	ExtensionType() int8  // -> uint8 	Len() int 	MarshalBinaryTo([]byte) error 	UnmarshalBinary([]byte) error   } 5. ype RawExtension struct { 	Data []byte 	Type int8   //-> uint8   }
AND Other related changes

+++ msgp/write.go
@@ -413,7 +413,7 @@ func (mw *Writer) WriteUint64(u uint64) error {
 }
 
 // WriteByte is analogous to WriteUint8
-func (mw *Writer) WriteByte(u byte) error { return mw.WriteUint8(uint8(u)) }
+func (mw *Writer) WriteByte(u byte) error { return mw.WriteUint8(u) }
 
 // WriteUint8 writes a uint8 to the writer
 func (mw *Writer) WriteUint8(u uint8) error { return mw.WriteUint64(uint64(u)) }
